feat(cli): add EKS cluster detection to user kubeconfig generation
This change adds EKS cluster detection to the user kubeconfig generation logic.
It checks for the presence of the aws-auth ConfigMap in the kube-system namespace
to determine if the cluster is an EKS cluster. If it is, it generates a kubeconfig
that uses the aws-iam-authenticator to authenticate to the cluster. If it is not,
it generates a kubeconfig that uses a service account token.

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>
ghstack-source-id: 7c93ebc7a7e72604754621efe8ebfeabb1cf87cd
ghstack-comment-id: 3499290897
Pull-Request: https://github.com/seemethere/devservers/pull/117